feat: use pg_ready for k8s migration job init-check
nslookup of hostname was used as db init check for
migration job. This was causing the migration job
container to fail multiple times. NSLOOKUP being
succesfull does not equate to db being ready to accept
connections.

This PR changed that to use pg_ready tool which is much more
reliable metric to assess db ready status.