feat: use pg_ready for k8s migration job init-check (#34)
nslookup of DB hostname was used as DB init check for migration job. This was
causing the migration job container to fail multiple times. NSLOOKUP being 
successful does not equate to DB is ready to accept connections.

This PR changed that to use the pg_ready tool which is a much more reliable 
metric to assess DB ready status.